fix: Restore custom marshaler support for input data conversion
The previous refactoring inadvertently removed the ability to apply custom
marshalers to input data before jq processing. This fix restores that
functionality while maintaining the clean architecture.

Changes:
- Restore yaml.EncodeOption parameter to convertToJQCompatible
- Add encodeOptions field back to defaultInputMarshaler
- Use yamlformat.MarshalJSON for complex type conversion (not json.Marshal)
- Add comprehensive design documentation explaining the rationale
- Add tests verifying input data custom marshaler functionality

This ensures users can control how their custom types (e.g., time.Time)
are converted to jq-compatible format before query execution, which is
essential for many real-world use cases.

+++ public_api_test.go
@@ -80,7 +80,7 @@ func TestWithCallback(t *testing.T) {
 			name:          "error in callback",
 			input:         []int{1, 2, 3},
 			query:         ".[]",
-			expected:      []interface{}{float64(1), float64(2)},
+			expected:      []interface{}{1, 2},
 			wantErr:       true,
 			expectedError: errors.New("stop!"),
 		},
